+++ tests/test_cli.py
@@ -1,3 +1,5 @@
+import os
+
 import aqt
 
 
@@ -47,3 +49,31 @@ def test_cli_check_mirror():
     args = cli.parser.parse_args(arg)
     assert args.base == 'https://download.qt.io/'
     assert cli._check_mirror(args.base)
+
+
+def test_cli_make_config_file(tmp_path):
+    cli = aqt.cli.Cli()
+    os.makedirs(tmp_path.joinpath('5.11.3', 'clang_64', 'bin'))
+    os.makedirs(tmp_path.joinpath('5.11.3', 'clang_64', 'mkspecs'))
+    with tmp_path.joinpath('5.11.3', 'clang_64', 'mkspecs', 'qconfig.pri').open('w') as f:
+        f.write("QT_ARCH = x86_64\n"
+                "QT_BUILDABI = x86_64-little_endian-lp64\n"
+                "QT.global.enabled_features = shared rpath c++11 c++14 future "
+                "concurrent pkg-config separate_debug_info\n"
+                "QT.global.disabled_features = cross_compile framework appstore-compliant debug_and_release "
+                "simulator_and_device build_all c++1z force_asserts static\n"
+                "QT_CONFIG += shared rpath release c++11 c++14 concurrent dbus "
+                "reduce_exports reduce_relocations separate_debug_info stl\n"
+                "CONFIG += shared release\n"
+                "QT_VERSION = 5.11.3\n"
+                "QT_MAJOR_VERSION = 5\n"
+                "QT_MINOR_VERSION = 11\n"
+                "QT_PATCH_VERSION = 3\n"
+                "QT_GCC_MAJOR_VERSION = 5\n"
+                "QT_GCC_MINOR_VERSION = 3\n"
+                "QT_GCC_PATCH_VERSION = 1\n"
+                "QT_EDITION = Enterprise\n"
+                "QT_LICHECK = licheck64\n"
+                "QT_RELEASE_DATE = 2018-11-29\n")
+    cli.make_conf_files('5.11.3', 'clang_64', tmp_path)
+    assert os.path.exists(tmp_path.joinpath('5.11.3', 'clang_64', 'bin', 'qt.conf'))